kodemy-auth/src/main/java/pl/sknikod/kodemyauth/util/data/AuditorAwareAdapter.java
package pl.sknikod.kodemyauth.util.data;

import org.springframework.data.domain.AuditorAware;
import org.springframework.lang.NonNull;
import org.springframework.security.core.Authentication;
import org.springframework.security.core.context.SecurityContextHolder;
import pl.sknikod.kodemycommon.security.UserPrincipal;

import java.util.Optional;

public class AuditorAwareAdapter implements AuditorAware<Long> {

    @Override
    @NonNull
    public Optional<Long> getCurrentAuditor() {
        return Optional.ofNullable(SecurityContextHolder.getContext().getAuthentication())
                .filter(Authentication::isAuthenticated)
                .map(Authentication::getPrincipal)
                .filter(UserPrincipal.class::isInstance)
                .map(UserPrincipal.class::cast)
                .map(UserPrincipal::getId)
                .or(Optional::empty);
    }
}
